zapisywanie ukladu jako stringi bez rekurencji
